refactored into higher level directory files. Renamed the test script Play.py, which still works so Im assuming all is good?
