+++ game_files/__init__.py
@@ -1,0 +1,1 @@
+from .game import Game
